requirements.txt
discord.py==2.3.2
pendulum==2.1.2
psutil==5.9.5
rich==13.6.0
SQLAlchemy==2.0.21
watchdog==3.0.0
